Copy skip memory mgmnt enter/exit options from mpfavorites to mplogin & mpdata
The mpfavorites utility already has the -sme and -smx options to allow
the entry and exiting of memory management mode to be skipped.  This
change copies those options to the mplogin and mpdata utilities so
that multiple commands that require memory management mode can be
performed without having to go through the hassle of entering your PIN
for each command.