Prompt user for password before opening Mooltipass
This change was prompted by trying to move a Mooltipass password to a
different context by using the Mooltipass to enter the password When
mplogin prompted "Enter the password to use:".  If you tried to do
this before this change, the message "HEY I GOT A 0xC4!" would be
printed by recv_packet() in mooltipass.py.  I don't see any reason for
mplogin to delay calling getpass.getpass() until after the Mooltipass
was opened so I moved it to be a part of the argparse handling in
mplogin's main_options() to avoid the 0xC4 problem.