Remove obsolete Python constructs
The '<>' operator is considered obsolete in favor of the '!=' operator
per https://docs.python.org/2.7/reference/lexical_analysis.html and in
Python 3 it actually is an error.  There was a single instance of '<>'
in mooltipass_client.py which is changed to '!='.

The other change was to add parentheses around the handled exceptions
on 'except' statements in mpdata.py and mplogin.py.  Although not
officially obsolete, it is considered the old way to do things per
https://docs.python.org/2/tutorial/errors.html and also results in an
error in Python 3.