Remove code that handles -sme option now that it is automatic.
Now that entering memory management mode will automatically be skipped
if the Mooltipy is already in memory management mode, there is no need
for the code to skip mooltipass.start_memory_management().  This
change does not remove the parsing of the -sme option though in case
it is desired to leave it for backwards compatibility.