Python 3: Accomodate str/bytes changes
Changing from Python 2 to 3 causes various properties and methods
of mooltipy to be of the bytes type instead of str.  Add code to
change them back to str so their users don't all have to change.

In addition, there are some additional instances where arrays of bytes
need to be initialized with bytes instead of strings.