Python 3: Add required=True parameter to add_subparsers().
If a valid option (e.g. -smx), but no command (i.e. set, get, list or
del) was provided to mpfavorites.py, it would give a KeyError
traceback.  mpdata.py and mplogin.py would print:
  An error occurred:
  None

I'm not completely sure that this came about as a result of moving to
Python 3, but it can be remedied by adding required=True to the
add_subparsers() calls.  After that, they display something like:
  usage: mplogin.py [-h] ... {get,set,del,list} context
  mplogin.py: error: the following arguments are required: command